chore: ⛑ Reorganize routes
