+++ src/core/router/types.ts
@@ -1,0 +1,9 @@
+import type { RouteObject } from 'react-router-dom'
+
+export interface CustomRouteObject extends Omit<RouteObject, 'children' | 'path'> {
+  path?: string | string[]
+  private?: boolean
+  onlyPublic?: boolean
+  redirect?: string
+  children?: CustomRouteObject[]
+}
